CatalogTriggerInterpret provides a hardcoded list of triggers for Catalog Metadata Descriptor, so Catalog Consensus plugin is fired.

+++ java/muba-catalogs/src/main/java/com/wrupple/muba/catalogs/domain/CatalogEventListener.java
@@ -11,7 +11,7 @@ public interface CatalogEventListener extends UserDefinedCatalogJob {
 	String CATALOG = "CatalogEventListener";
 	String ACTION_FIELD="action";
 
-    Integer getAction();
+    Long getAction();
 
 	
 	/**
